refactor: Player들을 가지고 있는 Players 객체 추가 및 Dealer가 가지고 있던 이름들 유효성검증도 Players로 변경
